changes of busines logic of app
