+++ src/components/CalendarGrid/CalendarGrid.jsx
@@ -1,22 +1,17 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import moment from 'moment';
 import { useLocation } from 'react-router';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import classes from './CalendarGrid.module.css';
 import Day from '../Day/Day';
-import { checkDay } from '../../utils/checkers';
 
 import { v4 as uuidv4 } from 'uuid';
-import { ADD_SHIFT, REMOVE_SHIFT } from '../../reducers/shiftReducer';
-import { Context } from '../../context/Context';
 
 const CalendarGrid = function ({ startDay, endDay, today }) {
   const totalDays = [];
-  const dispatch = useDispatch();
   const day = startDay.clone();
   const startMonth = today.clone().startOf('month');
   const location = useLocation();
-  const { selectedEmployee } = useContext(Context);
   const shifts = useSelector((state) => {
     return location.pathname === '/moderation'
       ? state.shiftReducer.tempShifts
@@ -50,42 +45,10 @@ const CalendarGrid = function ({ startDay, endDay, today }) {
 
   calculateDays();
 
-  const addShift = (day) => {
-    const employeeIdArray = day.shifts.map((shift) => shift.employee.id);
-
-    const newShift = {
-      id: uuidv4(),
-      employee: selectedEmployee,
-      month: moment(day.fullday).format('M'),
-      day: moment(day.fullday).format('D'),
-    };
-
-    if (location.pathname === '/moderation') {
-      if (selectedEmployee) {
-        if (checkDay(day.fullday)) {
-          employeeIdArray.includes(selectedEmployee.id)
-            ? alert('Невозможно добавить две смены одному специалисту в один день!')
-            : dispatch({ type: ADD_SHIFT, shift: newShift });
-          return;
-        }
-        alert('Невозможно произвести операцию задним числом!');
-        return;
-      }
-      alert('Выберите сотрудника для добавления в график!');
-    }
-  };
-
-  const removeShift = (shift) => {
-    dispatch({
-      type: REMOVE_SHIFT,
-      shift: shift,
-    });
-  };
-
   return (
     <div className={classes.wrapper}>
       {totalDays.map((dayObj) => (
-        <Day day={dayObj} addShift={addShift} key={dayObj.id} removeShift={removeShift} />
+        <Day day={dayObj} key={dayObj.id} />
       ))}
     </div>
   );
